[WebAppProvider] Add is_uninstalling to database, cleanup on startup.
It is possible for uninstalls to be cut off before they complete if the
browser is shut down or if there is a crash, leaving them in a semi-
broken state. This change saves the 'uninstalling' state in the
database, and retries the uninstall on startup.

Bug: 1162477, 1124047
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3035812
Auto-Submit: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Jiewei Qian  <qjw@chromium.org>
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#935725}

+++ chrome/browser/web_applications/web_app_uninstall_job.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "chrome/browser/web_applications/web_app_uninstall_commit_task.h"
+#include "chrome/browser/web_applications/web_app_uninstall_job.h"
 
 #include "base/bind.h"
 #include "base/logging.h"
@@ -17,7 +17,7 @@
 
 namespace web_app {
 
-WebAppUninstallCommitTask::WebAppUninstallCommitTask(
+WebAppUninstallJob::WebAppUninstallJob(
     OsIntegrationManager* os_integration_manager,
     WebAppSyncBridge* sync_bridge,
     WebAppIconManager* icon_manager,
@@ -29,23 +29,25 @@ WebAppUninstallCommitTask::WebAppUninstallCommitTask(
       registrar_(registrar),
       profile_prefs_(profile_prefs) {}
 
-WebAppUninstallCommitTask::~WebAppUninstallCommitTask() = default;
+WebAppUninstallJob::~WebAppUninstallJob() = default;
 
-void WebAppUninstallCommitTask::Start(const AppId& app_id,
-                                      url::Origin app_origin,
-                                      webapps::WebappUninstallSource source,
-                                      UninstallCallback callback) {
+void WebAppUninstallJob::Start(const AppId& app_id,
+                               url::Origin app_origin,
+                               webapps::WebappUninstallSource source,
+                               ModifyAppRegistry delete_option,
+                               UninstallCallback callback) {
   app_id_ = app_id;
   source_ = source;
+  delete_option_ = delete_option;
   callback_ = std::move(callback);
   DCHECK(state_ == State::kNotStarted);
-  state_ = State::kPendingOsHooksDeletion;
+  state_ = State::kPendingDataDeletion;
 
   // Note: It is supported to re-start an uninstall on startup, so
   // `is_uninstalling()` is not checked. It is a class invariant that there can
   // never be more than one uninstall task operating on the same web app at the
   // same time.
-  {
+  if (delete_option_ == ModifyAppRegistry::kYes) {
     ScopedRegistryUpdate update(sync_bridge_);
     WebApp* app = update->UpdateApp(app_id);
     DCHECK(app);
@@ -56,40 +58,56 @@ void WebAppUninstallCommitTask::Start(const AppId& app_id,
   RemoveAppIsolationState(profile_prefs_, app_origin);
 
   os_integration_manager_->UninstallAllOsHooks(
-      app_id_, base::BindOnce(&WebAppUninstallCommitTask::OnOsHooksUninstalled,
+      app_id_, base::BindOnce(&WebAppUninstallJob::OnOsHooksUninstalled,
                               weak_ptr_factory_.GetWeakPtr()));
+  icon_manager_->DeleteData(
+      app_id, base::BindOnce(&WebAppUninstallJob::OnIconDataDeleted,
+                             weak_ptr_factory_.GetWeakPtr()));
 }
 
-void WebAppUninstallCommitTask::OnOsHooksUninstalled(OsHooksErrors errors) {
-  DCHECK(state_ == State::kPendingOsHooksDeletion);
+void WebAppUninstallJob::StopAppRegistryModification() {
+  delete_option_ = ModifyAppRegistry::kNo;
+}
+
+void WebAppUninstallJob::OnOsHooksUninstalled(OsHooksErrors errors) {
+  DCHECK(state_ == State::kPendingDataDeletion);
+  hooks_uninstalled_ = true;
   base::UmaHistogramBoolean("WebApp.Uninstall.OsHookSuccess", errors.none());
   errors_ = errors_ || errors.any();
-
-  {
-    DCHECK_NE(registrar_->GetAppById(app_id_), nullptr);
-    ScopedRegistryUpdate update(sync_bridge_);
-    update->DeleteApp(app_id_);
-  }
-  state_ = State::kPendingIconDataDeletion;
-  icon_manager_->DeleteData(
-      app_id_, base::BindOnce(&WebAppUninstallCommitTask::OnIconDataDeleted,
-                              weak_ptr_factory_.GetWeakPtr()));
+  MaybeFinishUninstall();
 }
 
-void WebAppUninstallCommitTask::OnIconDataDeleted(bool success) {
-  DCHECK(state_ == State::kPendingIconDataDeletion);
+void WebAppUninstallJob::OnIconDataDeleted(bool success) {
+  DCHECK(state_ == State::kPendingDataDeletion);
+  app_data_deleted_ = true;
   base::UmaHistogramBoolean("WebApp.Uninstall.IconDataSuccess", success);
   errors_ = errors_ || !success;
+  MaybeFinishUninstall();
+}
 
+void WebAppUninstallJob::MaybeFinishUninstall() {
+  DCHECK(state_ == State::kPendingDataDeletion);
+  if (!hooks_uninstalled_ || !app_data_deleted_)
+    return;
   state_ = State::kDone;
 
   base::UmaHistogramBoolean("WebApp.Uninstall.Result", !errors_);
 
   webapps::InstallableMetrics::TrackUninstallEvent(source_);
 
+  switch (delete_option_) {
+    case ModifyAppRegistry::kYes: {
+      DCHECK_NE(registrar_->GetAppById(app_id_), nullptr);
+      ScopedRegistryUpdate update(sync_bridge_);
+      update->DeleteApp(app_id_);
+    } break;
+    case ModifyAppRegistry::kNo:
+      DCHECK_EQ(registrar_->GetAppById(app_id_), nullptr);
+      break;
+  }
   registrar_->NotifyWebAppUninstalled(app_id_);
-  std::move(callback_).Run(errors_ ? WebAppUninstallCommitTaskResult::kError
-                                   : WebAppUninstallCommitTaskResult::kSuccess);
+  std::move(callback_).Run(errors_ ? WebAppUninstallJobResult::kError
+                                   : WebAppUninstallJobResult::kSuccess);
 }
 
 }  // namespace web_app
